refactor(ai): consolidate remaining root test scaffolding into test/support/ (#3717)
Move the last three flat AI test scaffolding helpers
(domain_planner_test_fake_api.dart, planner_test_helpers.dart,
growth_stage_planner_test_support.dart) from the test/ root into
test/support/, completing the "single coherent test/support/ mirror"
target state for #3717. Repoint all 56 importers under test/planning/
from ../<helper>.dart to ../support/<helper>.dart; the support-local
self-import stays same-dir.

Behaviour-neutral (test layout only): no lib changes. Both AI gates
(repo.ai_test_no_duplicate_scaffolding, repo.ai_test_mirrors_lib) stay
green since the moved files are non-_test.dart helpers.

Refs #3717
