+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_civilian_work_log_empty_candidates_test.dart
@@ -3,8 +3,8 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:logger/logger.dart';
 
-import '../domain_planner_test_fake_api.dart';
-import '../planner_test_helpers.dart';
+import '../support/domain_planner_test_fake_api.dart';
+import '../support/planner_test_helpers.dart';
 
 void main() {
   group('runDomainPlanners civilian work logging — empty workCandidates', () {
