+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_colonial_civilian_work_test.dart
@@ -5,7 +5,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../domain_planner_test_fake_api.dart';
+import '../support/domain_planner_test_fake_api.dart';
 
 /// Colonial civilian work through [runDomainPlanners] (Refs #2509).
 void main() {
