+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_develop_nw_purchase_suppression_test.dart
@@ -52,7 +52,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../domain_planner_test_fake_api.dart';
+import '../support/domain_planner_test_fake_api.dart';
 
 const String _nationId = 'gp1';
 const String _owProvince = 'oldWorld|home';
