+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_expand_gp_only_blocker_declare_test.dart
@@ -74,7 +74,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../domain_planner_test_fake_api.dart';
+import '../support/domain_planner_test_fake_api.dart';
 
 const String _nationId = 'gp1';
 const String _blockerGpId = 'gp2';
