+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_expand_minor_declare_war_test.dart
@@ -58,7 +58,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../domain_planner_test_fake_api.dart';
+import '../support/domain_planner_test_fake_api.dart';
 
 const String _nationId = 'gp1';
 const String _minorId = 'minor1';
