+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_expand_nw_work_suppression_test.dart
@@ -36,7 +36,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../domain_planner_test_fake_api.dart';
+import '../support/domain_planner_test_fake_api.dart';
 
 const String _nationId = 'gp1';
 const String _fieldArmyId = 'field_a';
