+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_move_orders_test.dart
@@ -3,7 +3,7 @@ import 'package:colonizethis_ai/colonizethis_ai.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
-import '../planner_test_helpers.dart';
+import '../support/planner_test_helpers.dart';
 
 void main() {
   group('runDomainPlanners', () {
