+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_move_work_xor_test.dart
@@ -3,7 +3,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../planner_test_helpers.dart';
+import '../support/planner_test_helpers.dart';
 
 /// [DefaultOrderSuggestionAPI] with fixed civilian work rows so Full AI assigns
 /// work first; move suggestions still use production [suggestMoveOrders] (draft
