+++ packages/colonizethis_ai/test/planning/growth_stage_planner_relocation_test.dart
@@ -12,8 +12,8 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import '../domain_planner_test_fake_api.dart';
-import '../growth_stage_planner_test_support.dart';
+import '../support/domain_planner_test_fake_api.dart';
+import '../support/growth_stage_planner_test_support.dart';
 
 void main() {
   group('growth-stage Builder relocation — AC7 feedstock province', () {
