+++ packages/colonizethis_ai/test/planning/phase_planner_conquest_frontier_march_test.dart
@@ -11,8 +11,8 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import '../domain_planner_test_fake_api.dart';
-import '../planner_test_helpers.dart';
+import '../support/domain_planner_test_fake_api.dart';
+import '../support/planner_test_helpers.dart';
 
 const ExpandMilitaryPlan _expandOwOnly = ExpandMilitaryPlan(
   priorityDestinationProvinceIdsSorted: <String>['oldWorld|minor1_a'],
