+++ packages/colonizethis_ai/test/planning/phase_planner_economy_first_naval_transport_bootstrap_test.dart
@@ -16,7 +16,7 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import '../domain_planner_test_fake_api.dart';
+import '../support/domain_planner_test_fake_api.dart';
 
 const String _nationId = 'gp3';
 const String _owHome = 'oldWorld|gp3_0';
