+++ packages/colonizethis_ai/test/planning/planner_context_test.dart
@@ -4,7 +4,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import '../planner_test_helpers.dart';
+import '../support/planner_test_helpers.dart';
 
 void main() {
   group('PlannerContext weight resolution (Refs #2521)', () {
