+++ packages/colonizethis_ai/test/planning/recruitment_planner_test.dart
@@ -9,7 +9,7 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import '../domain_planner_test_fake_api.dart';
+import '../support/domain_planner_test_fake_api.dart';
 
 const _config = AIConfig(
   leaderId: 'victoria',
