+++ packages/colonizethis_ai/test/planning/research_planner_multi_slot_test.dart
@@ -5,8 +5,8 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import '../domain_planner_test_fake_api.dart';
-import '../planner_test_helpers.dart';
+import '../support/domain_planner_test_fake_api.dart';
+import '../support/planner_test_helpers.dart';
 
 /// Treasury-aware multi-slot funding for the Full-AI research planner
 /// (Refs #3472). Uses a fake suggestion API so the funding/packing/preserve
